requirements.txt
bcrypt
passlib
absl-py
asn1crypto
astor
astroid
certifi
chardet
Click
cryptography
decorator
defer
ecdsa
editdistance~=0.6.0
Flask~=2.0.2
Flask-Cors
flask-httpauth
gast
h5py
httplib2
idna
ipython
ipython-genutils
isort
itsdangerous
jedi
Jinja2
jws
Keras-Applications
Keras-Preprocessing
keyring
keyrings.alt
lazy-object-proxy
Markdown
MarkupSafe
mccabe
mongoengine~=0.23.1
numpy~=1.21.2
olefile
opencv-python~=4.5.3.56
parso
pexpect
pickleshare
Pillow~=8.3.1
plumbum
prompt-toolkit
Pygments
pylint
pymongo
pyxdg
requests~=2.25.1
requests-unixsocket
SecretStorage
simplegeneric
six
setuptools
tensorboard
tensorflow~=2.7.0
termcolor
traitlets
typed-ast
urllib3
waitress
wcwidth
websocket-client
Werkzeug
wrapt
grpcio
protobuf

matplotlib~=3.4.2